move tests into __tests__ folder
