+++ src/api/__test__/trackerTrackedEntities.integration.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeAll, expectTypeOf } from "vitest";
-import { getPlayD2Api, playFixtures } from "../testing";
-import { D2Api } from "../2.42";
+import { getPlayD2Api, playFixtures } from "../../testing";
+import { D2Api } from "../../2.42";
 
 describe("TrackedEntities.get against play.dhis2.org/42", () => {
     let api: D2Api;
